Moved component files to folder gamesList
