+++ frontend/components/gamesList/GameListRow.tsx
